+++ src/lib/contexts/channels/channel-provider.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { getContext, onMount, setContext, type Snippet } from 'svelte';
   import { channelContext } from './channel-context.svelte';
-  import type { MyUserContext } from './my-user-context.svelte';
+  import type { MyUserContext } from '../users/my-user-context.svelte';
 
   const myUserContext = getContext<MyUserContext>('myUserContext');
