+++ src/routes/chat/components/chat-list.svelte
@@ -3,8 +3,8 @@
   import { formatDistanceToNow } from 'date-fns';
   import type { ChannelListItem } from '@baragaun/bg-node-client';
   import ChannelOptionsMenu from './channel-options-menu.svelte';
-  import { myUserContext } from '@/contexts/my-user-context.svelte';
-  import { channelContext } from '@/contexts/channel-context.svelte';
+  import { myUserContext } from '@/contexts/users/my-user-context.svelte';
+  import { channelContext } from '@/contexts/channels/channel-context.svelte';
 
   let { channels }: { channels: ChannelListItem[] } = $props();
