+++ src/routes/settings/account/components/delete-account-form.svelte
@@ -7,7 +7,7 @@
   import FormButton from '@/components/forms/form-button.svelte';
   import IdentFormInput from '@/components/forms/form-ident-input.svelte';
   import { Button } from '@/components/ui/button';
-  import { MyUserContext, myUserContext } from '@/contexts/my-user-context.svelte';
+  import { MyUserContext, myUserContext } from '@/contexts/users/my-user-context.svelte';
   import translate from '@/helpers/language/translate';
   import { m } from '@/paraglide/messages';
   import { AppUiMessage } from '@/types/enums';
