+++ src/routes/settings/account/components/update-email-form.svelte
@@ -5,7 +5,7 @@
   import { Button } from '@/components/ui/button';
   import { Input } from '@/components/ui/input';
   import { MsaListenerHandler } from '@/contexts/msa-listener-handler.svelte';
-  import { MyUserContext } from '@/contexts/my-user-context.svelte';
+  import { MyUserContext } from '@/contexts/users/my-user-context.svelte';
   import translate from '@/helpers/language/translate';
   import { m } from '@/paraglide/messages';
   import { AppUiMessage } from '@/types/enums';
